move scripts to dedicated src folder
