+++ analysis_utils/binning_utils.py
@@ -5,9 +5,9 @@ Utilities for values binning
 
 import pandas as pd
 
-from cochange_analysis import the_lower_the_better
-from feature_pair_analysis import pair_features_vs_concept, features_stats_to_cm_df
-from ml_utils import extract_relevent_features
+from analysis_utils.cochange_analysis import the_lower_the_better
+from analysis_utils.feature_pair_analysis import pair_features_vs_concept, features_stats_to_cm_df
+from analysis_utils.ml_utils import extract_relevent_features
 
 SIDES_SUFFIX = '_SIDES'
